set `ignore_angle=True` as default.
The warning already suggests that this is what it does
```
This results in lose of keypoint, segmentation and bbox rotation information if present
```

So it makes sense that ignoring angle is the default behaviour.

Also, if the `self._ignore_angle=False` and a detection has a non-zero angle, it crashes the pipeline. So it's better when the default behaviour doesn't crash a pipeline

- improve the error message when `_ignore_angle=False` and a detection has a non-zero rotation angle